Move http api blueprint outside of plugins
